backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# Deep Learning
torch==2.1.2
torchvision==0.16.2
ultralytics==8.1.24
transformers==4.37.0

# Tracking & ML Utils
filterpy==1.4.5
scikit-learn==1.4.0
scipy==1.11.4

# Search & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# Utils
pydantic==2.5.3
python-dotenv==1.0.0
aiofiles==23.2.1
scikit-image==0.22.0

# Dataset download
datasets==2.16.1
tqdm==4.66.1
